% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFinder.R
\name{CFinder}
\alias{CFinder}
\title{Identify potential confounding factors or batch effects}
\usage{
CFinder(data, pheno, type)
}
\arguments{
\item{data}{A matrix, the normalized gene/microRNA expression dataset, should be a numeric matrix, with rows referring to genes/microRNAs and columns to samples.}

\item{pheno}{A vector of sample phenotypes. Sample phenotype in a scientific research could be categorical (treatment/control), or continuous (age). If you have multiple phenotypes, use a list with names denoting the corresponding phenotype.}

\item{type}{A string, 'categorical' if sample phenotype if categorical, 'continuous' if sample phenotype is continuous.}
}
\value{
A vector containing p-values for top k significant PCs.
}
\description{
\code{CFinder} first does a principal component analysis via singular value decomposition, and compute correlation of a significant PCs (dimensionality estimated by \code{\link[isva]{EstDimRM}}) with sample class information.
}
\details{
This function computes the moderated t-statistic for users using empirical Bayes method, it is especially useful when the sample size is too small to perform parametric tests.

Given a normalized gene/microRNA expression data matrix and a vector indicating sample phenotype/class, \code{CFinder} first centers the input data matrix, then estimate the number of significant PCs using \code{\link[isva]{EstDimRM}}. Next, it does a singular value decomposition to the data matrix. Then it computes the correlation of significant PCs with potential confounding factors by using linear correlation (for continueous phenotype data like age) or Kruskal-Wallis rank sum test (for categorical phenotype data like tumor subtype).
}
\examples{
# prepare your normalized data matrix
data.m <- matrix(rnorm(120), nrow = 20, ncol = 6)
# prepare the phenotype info ("C"-control; "T"-treatment)
class.v <- c('C', 'C', 'C', 'T', 'T', 'T')
# run function
p.v <- CFinder(data = data.m, pheno = class.v, type = 'categorical')
}
\seealso{
\code{\link[isva]{EstDimRM}} for estimating the dimensionality of a dataset.
}
