% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairComp.R
\name{pairComp}
\alias{pairComp}
\title{Do differential gene/microRNA expression analysis}
\usage{
pairComp(data, class, padj = "fdr")
}
\arguments{
\item{data}{A matrix, the normalized gene/microRNA expression dataset, should be a numeric matrix, with rows referring to genes/microRNAs and columns to samples.}

\item{class}{A vector of sample phenotypes. Sample phenotype in a scientific research could be treatment/control, normal/cancer or smoker/non-smoker. Different phenotypes could each be encoded as 0/1 when inputting to \code{LimmaFn}, for example, Normal-0; Cancer-1.}
}
\value{
A table with rows for all genes (ranked by significance) and columns of log2 fold-change, average expression, moderated t-statistic, p-value, adjusted p-value (default to Benjamini–Hochberg procedure). The table is the output of \code{\link[limma]{topTable}} function.
}
\description{
\code{pairComp} uses functions in \code{limma} package to easily compute the moderated t-statistics and p-values from differential gene/microRNA expression tests comparing between different phenotypes even when sample size is small.
}
\details{
This function computes the moderated t-statistic for users using empirical Bayes method, it is especially useful when the sample size is too small to perform parametric tests.

Given a normalized gene expression or DNA methylation data matrix and a vector indicating sample phenotype, \code{LimmaFn} first fits a linear model using \code{\link[limma]{lmFit}}, then it refits the model and do comparisons between any two different phenotypes with \code{\link[limma]{contrasts.fit}}, finally it estimates moderated t-statistics for each comparison from the fitted model using empirical Bayes method (\code{\link[limma]{eBayes}}) and output the result from the \code{\link[limma]{topTable}} function.

Note that doing the \code{contrasts.fit} step will not make a difference if you do comparison between two different sample status (treatment/control). However, When there are more than two sample status in your data set, this step will do comparison between every two status. And resulted summary tables will be stored in a list.
}
\examples{
# prepare your normalized data matrix
data.m <- matrix(rnorm(120), nrow = 20, ncol = 6)
# prepare the phenotype info ("C"-control; "T"-treatment)
class.v <- c('C', 'C', 'C', 'T', 'T', 'T')
# run function
lim.o <- pairComp(data = data.m, class = class.v)
}
\seealso{
\code{\link[limma]{lmFit}} for fitting a linear model, \code{\link[limma]{contrasts.fit}} for refitting, \code{\link[limma]{eBayes}} for Bayes method, \code{\link[limma]{topTable}} for the output table.
}
